requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
sentencepiece>=0.1.99
sacrebleu>=2.3.0
gradio>=4.0.0
langdetect>=1.0.9
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
